fix: correct all page.css import paths after moving file
Updated all imports of page.css to use correct relative paths based on new location in dashboard/components.
Fixed "Module not found" errors caused by incorrect or outdated import paths.
Ensured consistent background animation styling across dashboard and user pages.
