--- src/app/page.css
@@ -1,37 +1,0 @@
-/* Background animation container */
-.animated-lines {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 200%;
-    height: 200%;
-    background-image: linear-gradient(to right, transparent 98%, #e0e3e9 98%, #b3b9b9 100%), 
-                      linear-gradient(to bottom, transparent 98%, #fdfeff 98%, #a8aeb1 100%);
-    background-size: 50px 50px;
-    animation: move-lines 4s linear infinite; 
-    z-index: 0;
-  }
-  
-
-  @keyframes move-lines {
-    0%{
-      transform: translate(0, 0);
-      opacity:0.5;
-    }
-    25% {
-      transform: translate(5px, 5px); 
-      opacity:0.75;
-    }
-    50% {
-        transform: translate(10px, 10px); 
-        opacity:1;
-      }
-      75% {
-        transform: translate(40px, 40px); 
-        opacity:0.75;
-      }
-      100% {
-        transform: translate(50px, 50px); 
-        opacity:0.5;
-      }
-  }
