+++ src/app/user/login/page.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import InputBox from "../components/InputBox";
 import Button from "../components/SubmitButton";
-import '../../page.css';
+import '../../dashboard/components/page.css';
 
 export default function Page() {
   return (
